requirements.txt
httpx[http2]
python-dotenv
fire

fastapi
uvicorn

pygame
imgui[pygame]